+++ src/Strategy/Maven/Plugin.hs
@@ -17,10 +17,11 @@ module Strategy.Maven.Plugin (
   textArtifactToPluginOutput,
   execPluginAggregate,
   execPluginReactor,
+  mavenCmdCandidates,
 ) where
 
 import Control.Algebra (Has)
-import Control.Effect.Diagnostics (Diagnostics, warn)
+import Control.Effect.Diagnostics (Diagnostics, recover, warn)
 import Control.Effect.Exception (Lift, bracket)
 import Control.Effect.Lift (sendIO)
 import Control.Monad (when)
@@ -30,6 +31,8 @@ import Data.ByteString qualified as BS
 import Data.FileEmbed.Extra (embedFile')
 import Data.Foldable (Foldable (fold), foldl')
 import Data.Functor (void)
+import Data.List.NonEmpty (NonEmpty (..))
+import Data.List.NonEmpty qualified as NE
 import Data.Map (Map)
 import Data.Map qualified as Map
 import Data.Maybe (catMaybes, isNothing)
@@ -37,11 +40,16 @@ import Data.String.Conversion (toText)
 import Data.Text (Text)
 import Data.Traversable (for)
 import Data.Tree (Tree (..))
+import DepTypes (DepType (MavenType))
+import Discovery.Walk (findFileInAncestor)
 import Effect.Exec (
   AllowErr (Never),
+  CandidateAnalysisCommands (..),
+  CandidateCommandEffs,
   Command (..),
   Exec,
   execThrow,
+  mkAnalysisCommand,
  )
 import Effect.ReadFS (
   ReadFS,
@@ -62,6 +70,7 @@ import Path (
 import Path.IO (createTempDir, getTempDir, removeDirRecur)
 import Strategy.Maven.PluginTree (TextArtifact (..), parseTextArtifact)
 import System.FilePath qualified as FP
+import System.Info qualified as SysInfo
 
 data DepGraphPlugin = DepGraphPlugin
   { group :: Text
@@ -107,17 +116,23 @@ withUnpackedPlugin plugin act =
 
       act pluginJarFilepath
 
-installPlugin :: (Has Exec sig m, Has Diagnostics sig m) => Path Abs Dir -> FP.FilePath -> DepGraphPlugin -> m ()
-installPlugin dir path plugin = void $ execThrow dir (mavenInstallPluginCmd path plugin)
+installPlugin :: (CandidateCommandEffs sig m, Has ReadFS sig m) => Path Abs Dir -> FP.FilePath -> DepGraphPlugin -> m ()
+installPlugin dir path plugin = do
+  cmd <- mavenInstallPluginCmd dir path plugin
+  void $ execThrow dir cmd
 
 execPlugin :: (Has Exec sig m, Has Diagnostics sig m) => (DepGraphPlugin -> Command) -> Path Abs Dir -> DepGraphPlugin -> m ()
 execPlugin pluginToCmd dir plugin = void $ execThrow dir $ pluginToCmd plugin
 
-execPluginAggregate :: (Has Exec sig m, Has Diagnostics sig m) => Path Abs Dir -> DepGraphPlugin -> m ()
-execPluginAggregate = execPlugin mavenPluginDependenciesCmd
+execPluginAggregate :: (CandidateCommandEffs sig m, Has ReadFS sig m) => Path Abs Dir -> DepGraphPlugin -> m ()
+execPluginAggregate dir plugin = do
+  cmd <- mavenPluginDependenciesCmd dir plugin
+  execPlugin (const cmd) dir plugin
 
-execPluginReactor :: (Has Exec sig m, Has Diagnostics sig m) => Path Abs Dir -> DepGraphPlugin -> m ()
-execPluginReactor = execPlugin mavenPluginReactorCmd
+execPluginReactor :: (CandidateCommandEffs sig m, Has ReadFS sig m) => Path Abs Dir -> DepGraphPlugin -> m ()
+execPluginReactor dir plugin = do
+  cmd <- mavenPluginReactorCmd dir plugin
+  execPlugin (const cmd) dir plugin
 
 outputFile :: Path Rel File
 outputFile = $(mkRelFile "target/dependency-graph.txt")
@@ -191,58 +206,75 @@ textArtifactToPluginOutput
                   , outEdges = newEdges <> cEdges
                   }
 
-mavenInstallPluginCmd :: FP.FilePath -> DepGraphPlugin -> Command
-mavenInstallPluginCmd pluginFilePath plugin =
-  Command
-    { cmdName = "mvn"
-    , cmdArgs =
-        [ "org.apache.maven.plugins:maven-install-plugin:3.0.0-M1:install-file"
-        , "-DgroupId=" <> group plugin
-        , "-DartifactId=" <> artifact plugin
-        , "-Dversion=" <> version plugin
-        , "-Dpackaging=jar"
-        , "-Dfile=" <> toText pluginFilePath
-        ]
-    , cmdAllowErr = Never
-    }
+-- | Search for maven wrappers in the closest parent directory and if found use them as a candidate command.
+mavenCmdCandidates :: (Has ReadFS sig m, Has Diagnostics sig m) => Path Abs Dir -> m CandidateAnalysisCommands
+mavenCmdCandidates dir =
+  mvnWrapperPath >>= \case
+    Just wrapper -> pure . mkCmd $ (toText wrapper) :| ["mvn"]
+    Nothing -> pure . mkCmd $ NE.singleton "mvn"
+  where
+    runningWindows :: Bool
+    runningWindows = SysInfo.os == "mingw32"
+    mkCmd :: NonEmpty Text -> CandidateAnalysisCommands
+    mkCmd cmds = CandidateAnalysisCommands cmds ["-v"] $ Just MavenType
+    -- Unlike with the gradle wrapper, it's not _expected_ for maven projects to use the maven wrapper.
+    -- Given that, don't warn on failure to find a wrapper; only warn if we find a wrapper and it fails to execute.
+    mvnWrapperPath :: (Has ReadFS sig m, Has Diagnostics sig m) => m (Maybe (Path Abs File))
+    mvnWrapperPath =
+      recover $
+        if runningWindows
+          then findFileInAncestor dir "mvnw.bat"
+          else findFileInAncestor dir "mvnw"
 
--- |The aggregate command is documented
--- [here.](https://ferstl.github.io/depgraph-maven-plugin/aggregate-mojo.html)
-mavenPluginDependenciesCmd :: DepGraphPlugin -> Command
-mavenPluginDependenciesCmd plugin =
-  Command
-    { cmdName = "mvn"
-    , cmdArgs =
-        [ group plugin <> ":" <> artifact plugin <> ":" <> version plugin <> ":aggregate"
-        , "-DgraphFormat=text"
-        , -- display deps that appear multiple times in different scopes as a single node
-          "-DmergeScopes"
-        , -- Don't omit edges for deps appearing in multiple places in the graph
-          "-DreduceEdges=false"
-        , "-DshowVersions=true"
-        , "-DshowGroupIds=true"
-        , -- Deps that are optional in the graph will be tagged optional in the cli's output
-          -- this does not exclude them from sourceUnits.
-          "-DshowOptional=true"
-        , -- Repeat transitive deps for packages that appear multiple times
-          "-DrepeatTransitiveDependenciesInTextGraph=true"
-        ]
-    , cmdAllowErr = Never
-    }
+mavenInstallPluginCmd :: (CandidateCommandEffs sig m, Has ReadFS sig m) => Path Abs Dir -> FP.FilePath -> DepGraphPlugin -> m Command
+mavenInstallPluginCmd workdir pluginFilePath plugin = do
+  candidates <- mavenCmdCandidates workdir
+  mkAnalysisCommand candidates workdir args Never
+  where
+    args =
+      [ "org.apache.maven.plugins:maven-install-plugin:3.0.0-M1:install-file"
+      , "-DgroupId=" <> group plugin
+      , "-DartifactId=" <> artifact plugin
+      , "-Dversion=" <> version plugin
+      , "-Dpackaging=jar"
+      , "-Dfile=" <> toText pluginFilePath
+      ]
 
--- |The reactor command is documented
--- [here.](https://ferstl.github.io/depgraph-maven-plugin/reactor-mojo.html)
-mavenPluginReactorCmd :: DepGraphPlugin -> Command
-mavenPluginReactorCmd plugin =
-  Command
-    { cmdName = "mvn"
-    , cmdArgs =
-        [ group plugin <> ":" <> artifact plugin <> ":" <> version plugin <> ":reactor"
-        , "-DgraphFormat=json"
-        , "-DoutputFileName=" <> toText reactorOutputFilename
-        ]
-    , cmdAllowErr = Never
-    }
+-- | The aggregate command is documented
+--  [here.](https://ferstl.github.io/depgraph-maven-plugin/aggregate-mojo.html)
+mavenPluginDependenciesCmd :: (CandidateCommandEffs sig m, Has ReadFS sig m) => Path Abs Dir -> DepGraphPlugin -> m Command
+mavenPluginDependenciesCmd workdir plugin = do
+  candidates <- mavenCmdCandidates workdir
+  mkAnalysisCommand candidates workdir args Never
+  where
+    args =
+      [ group plugin <> ":" <> artifact plugin <> ":" <> version plugin <> ":aggregate"
+      , "-DgraphFormat=text"
+      , -- display deps that appear multiple times in different scopes as a single node
+        "-DmergeScopes"
+      , -- Don't omit edges for deps appearing in multiple places in the graph
+        "-DreduceEdges=false"
+      , "-DshowVersions=true"
+      , "-DshowGroupIds=true"
+      , -- Deps that are optional in the graph will be tagged optional in the cli's output
+        -- this does not exclude them from sourceUnits.
+        "-DshowOptional=true"
+      , -- Repeat transitive deps for packages that appear multiple times
+        "-DrepeatTransitiveDependenciesInTextGraph=true"
+      ]
+
+-- | The reactor command is documented
+--  [here.](https://ferstl.github.io/depgraph-maven-plugin/reactor-mojo.html)
+mavenPluginReactorCmd :: (CandidateCommandEffs sig m, Has ReadFS sig m) => Path Abs Dir -> DepGraphPlugin -> m Command
+mavenPluginReactorCmd workdir plugin = do
+  candidates <- mavenCmdCandidates workdir
+  mkAnalysisCommand candidates workdir args Never
+  where
+    args =
+      [ group plugin <> ":" <> artifact plugin <> ":" <> version plugin <> ":reactor"
+      , "-DgraphFormat=json"
+      , "-DoutputFileName=" <> toText reactorOutputFilename
+      ]
 
 newtype ReactorArtifact = ReactorArtifact
   { reactorArtifactName :: Text