+++ Changelog.md
@@ -2,7 +2,10 @@
 
 <!-- - title: description ([#](https://github.com/fossas/fossa-cli/pull/#)) -->
 
-## Unreleased
+## v3.7.0
+
+- Support Maven wrapper (`mvnw`) usage in Maven projects, and user-provided binary overrides for Maven projects ([#1149](https://github.com/fossas/fossa-cli/pull/1149))
+  For more information, see the [Maven strategy documentation](./docs/references/strategies/languages/maven/maven.md).
 - Installation Script: Verify that the sha256sum of the downloaded archive matches the recorded one. ([#1158](https://github.com/fossas/fossa-cli/pull/1158))
 
 ## v3.6.18